requirements.txt
Flask
gunicorn
weasyprint
pandas
python-dotenv
whitenoise
pymongo
